+++ client/src/command/command.ts
@@ -1,13 +1,16 @@
 import { commands, ExtensionContext } from "vscode";
-import { changeTargetLanguage } from "./changeTargetLanguage";
+import { changeTargetLanguage, changeTranslateSource, toggleEnableHover, toggleMultiLineMerge } from "./changeTargetLanguage";
 import { replaceRange, replaceSelections } from "./replaceSelections";
 import { selectLastHover } from "./select";
 
 export function registerCommands(context: ExtensionContext) {
     context.subscriptions.push(
         commands.registerCommand('commentTranslate.select', selectLastHover),
-        commands.registerCommand('commentTranslate.replaceRange', replaceRange),
+        commands.registerCommand('commentTranslate._replaceRange', replaceRange),
         commands.registerCommand('commentTranslate.replaceSelections', replaceSelections),
+        commands.registerCommand('commentTranslate._toggleMultiLineMerge', toggleMultiLineMerge),
+        commands.registerCommand('commentTranslate._changeTranslateSource', changeTranslateSource),
+        commands.registerCommand('commentTranslate.toggleEnableHover', toggleEnableHover),
         commands.registerCommand('commentTranslate.changeTargetLanguage', changeTargetLanguage)
     );
 }