+++ client/src/languageFeature/hover.ts
@@ -1,4 +1,5 @@
 import { Hover, languages, MarkdownString, Range } from "vscode";
+import { getConfig } from "../configuration";
 import { client } from "../extension";
 
 export interface ITranslateHover {
@@ -24,13 +25,16 @@ export function registerHover(canLanguages:string[] = []) {
             });
             if(!res) return null;
             let {translatedText,translatedLink,humanizeText,range} = res;
+            const base64TranslatedText = Buffer.from(translatedText).toString('base64');
             const separator = ' &nbsp;&nbsp;|&nbsp;&nbsp; ';
-            let replace = `[$(replace)](command:commentTranslate.replaceRange?${encodeURIComponent(JSON.stringify({uri,range,text:translatedText}))} "Replace")`;
-            let combine = `[$(combine)](command:workbench.action.openWorkspaceSettings?${encodeURIComponent(JSON.stringify({query:'commentTranslate',openToSide:true}))} "Combine Multi Line")`;
+            let replace = `[$(replace)](command:commentTranslate._replaceRange?${encodeURIComponent(JSON.stringify({uri,range,text:base64TranslatedText}))} "Replace")`;
+
+            let multiLine = await getConfig<boolean>('multiLineMerge');
+            let combine = `[$(${multiLine?'selection':'remove'})](command:commentTranslate._toggleMultiLineMerge "Toggle Combine Multi Line")`;
 
             let setting = `[$(settings-gear)](command:workbench.action.openWorkspaceSettings?${encodeURIComponent(JSON.stringify({query:'commentTranslate'}))} "Open Comment Translate setting")`;
 
-            let translate = `[$(sync)](command:workbench.action.openWorkspaceSettings?${encodeURIComponent(JSON.stringify({query:'commentTranslate',openToSide:true}))} "Change translate source")`;
+            let translate = `[$(sync)](command:commentTranslate._changeTranslateSource "Change translate source")`;
 
             let header = new MarkdownString(`[Comment Translate]&nbsp;&nbsp;${replace}&nbsp;&nbsp;${combine}&nbsp;&nbsp;${setting}${separator}${translate}&nbsp;&nbsp;${translatedLink}`,true);
             header.isTrusted = true;